priorPosterior

Analyse prior and posterior node distributions from MCMCtree analysis

Usage

priorPosterior(
  MCMCPrior,
  MCMCPosterior = NULL,
  inputTree,
  return.density = FALSE,
  rootCalibration = NULL
)

Arguments

MCMCPrior

prior of MCMC file from MCMCtree analysis using data=0

MCMCPosterior

posterior of MCMC file from MCMCtree analysis using data

inputTree

phylogeny in MCMCtree format used in MCMCtree analysis

return.density

Logical. Whether to return the density or original values for the effective prior and posterior.

rootCalibration

= NULL If NULL, then behaves as by default. Alternatively, if a user has specified a root prior in the MCMCtreeR control file it can be added here as a vector in the form it would appear tree.

Value

list containing node estimates for each distribution

"prior" distribution of effective prior

"posterior" distribution of posterior

"specifiedPrior" distribution of specified prior

Author

Mark Puttick, Pascal Title

Examples

data(MCMCtree.output)
#> Warning: data set ‘MCMCtree.output’ not found
# priorPosterior(MCMCPrior, 
# MCMCPosterior=MCMCtree.output$MCMCtree.posterior, 
# path.to.input.tree)