Read MCMCtree output tree into R

Read MCMCtree output tree into R to produce time-scaled tree in APE format, and a table of the mean and 95

Usage

readMCMCtree(inputPhy, forceUltrametric = TRUE, from.file = TRUE)

Arguments

inputPhy

file directory of 'Figtree' output from MCMCtree

forceUltrametric

alters branch lengths at tips so tree is fully ultrametric (default = TRUE)

from.file

Logical. Read a tree from file or locally from within R?

Value

apePhy time-scaled output tree from MCMCtree in APE format

nodeAges mean and 95

Author

Mark Puttick

Examples

data(MCMCtree.output)
#> Warning: data set ‘MCMCtree.output’ not found
attach(MCMCtree.output)
#> The following objects are masked from MCMCtree.output (pos = 4):
#> 
#>     MCMCtree.phy, MCMCtree.posterior
#> The following objects are masked from MCMCtree.output (pos = 6):
#> 
#>     MCMCtree.phy, MCMCtree.posterior
# tree and node ages
readMCMCtree(MCMCtree.phy, from.file=FALSE)
#> $apePhy
#> 
#> Phylogenetic tree with 103 tips and 102 internal nodes.
#> 
#> Tip labels:
#>   Dioscorea_villosa, Colchicum_autumnale, Smilax_bona-nox, Sorghum_bicolor, Zea_mays, Oryza_sativa, ...
#> 
#> Rooted; includes branch lengths.
#> 
#> $nodeAges
#>         mean 95%_lower 95%_upper
#> 104 7.767503     6.651     9.296
#> 105 7.306332     6.337     8.634
#> 106 6.644847     5.906     7.652
#> 107 6.228787     5.607     7.012
#> 108 5.944306     5.403     6.633
#> 109 5.554283     5.118     6.145
#> 110 5.001146     4.766     5.149
#> 111 4.466497     4.367     4.509
#> 112 4.272271     4.131     4.405
#> 113 3.515712     3.311     3.651
#> 114 2.329351     2.068     2.467
#> 115 2.114190     1.853     2.341
#> 116 1.913382     1.671     2.174
#> 117 1.661844     1.474     1.893
#> 118 1.242075     1.151     1.285
#> 119 1.048606     0.926     1.175
#> 120 0.959314     0.847     1.091
#> 121 0.680444     0.365     0.913
#> 122 0.895020     0.767     1.029
#> 123 0.783066     0.624     0.930
#> 124 0.481127     0.319     0.653
#> 125 0.208725     0.090     0.386
#> 126 0.348893     0.194     0.536
#> 127 1.544019     1.390     1.759
#> 128 1.423139     1.282     1.642
#> 129 1.311295     1.190     1.538
#> 130 1.182058     1.109     1.388
#> 131 0.761427     0.513     1.050
#> 132 1.261988     1.210     1.286
#> 133 1.152143     1.047     1.226
#> 134 1.069097     0.949     1.168
#> 135 0.965804     0.834     1.089
#> 136 0.861304     0.698     1.001
#> 137 0.756563     0.556     0.920
#> 138 0.576249     0.352     0.782
#> 139 0.866342     0.689     1.005
#> 140 0.736447     0.503     0.913
#> 141 1.070807     0.943     1.170
#> 142 0.972898     0.837     1.089
#> 143 0.886580     0.741     1.015
#> 144 0.760147     0.583     0.908
#> 145 0.666530     0.483     0.834
#> 146 0.362568     0.186     0.573
#> 147 0.455096     0.229     0.727
#> 148 1.014444     0.782     1.180
#> 149 0.783065     0.472     1.019
#> 150 3.177954     3.084     3.372
#> 151 2.886741     2.660     3.182
#> 152 0.952532     0.426     1.932
#> 153 0.209907     0.067     0.511
#> 154 2.438058     1.701     3.048
#> 155 2.114159     1.425     2.791
#> 156 0.630082     0.186     1.268
#> 157 1.252099     0.835     1.883
#> 158 0.893094     0.507     1.322
#> 159 1.139416     0.794     1.627
#> 160 0.952772     0.631     1.329
#> 161 0.684231     0.364     1.029
#> 162 0.346178     0.135     0.684
#> 163 4.042667     3.923     4.196
#> 164 3.572424     3.215     3.938
#> 165 2.694457     1.069     3.563
#> 166 3.906610     3.848     4.040
#> 167 1.157836     0.498     2.476
#> 168 4.089600     3.927     4.322
#> 169 1.262264     0.751     2.345
#> 170 0.880515     0.427     1.586
#> 171 0.118091     0.041     0.278
#> 172 4.870626     4.632     5.065
#> 173 4.671717     4.397     4.916
#> 174 3.997939     3.494     4.493
#> 175 3.397905     2.994     3.957
#> 176 2.823961     2.685     3.177
#> 177 1.778637     0.972     2.746
#> 178 1.043410     0.699     1.538
#> 179 0.864497     0.545     1.262
#> 180 0.290700     0.161     0.498
#> 181 0.224755     0.098     0.429
#> 182 0.228450     0.104     0.426
#> 183 0.323686     0.126     0.642
#> 184 4.184525     4.053     4.461
#> 185 2.726547     2.283     3.621
#> 186 1.227128     0.707     2.209
#> 187 0.794258     0.332     1.346
#> 188 2.439996     0.792     3.922
#> 189 0.164268     0.053     0.382
#> 190 4.971913     4.089     5.802
#> 191 4.515646     3.554     5.409
#> 192 4.064539     2.936     5.021
#> 193 3.347874     2.004     4.445
#> 194 2.066850     0.860     3.533
#> 195 4.470035     3.452     5.384
#> 196 3.595206     2.097     4.690
#> 197 3.075275     1.413     4.250
#> 198 4.738018     3.116     6.056
#> 199 3.176621     1.125     4.842
#> 200 4.686395     2.547     6.542
#> 201 5.884835     3.952     7.700
#> 202 4.184454     2.113     6.224
#> 203 6.680455     5.025     8.412
#> 204 5.843148     4.500     7.406
#> 205 4.869119     3.227     6.683
#> 
#> attr(,"class")
#> [1] "MCMCtreer"