Find Descendent Tips From A Common Ancestor

This function finds tip descendants from a common ancestor

Usage

tipDes(phy, node)

Arguments

phy

user tree in ape format

node

one or more nodes from the ape format that designate the crown monophyletic group

Details

If a single node number is supplied the function returns a vector of tip labels. When more than one node number is supplied the function returns a list with each element a vector of tip labels for that node.

Author

Mark Puttick

Examples

set.seed(1029)
# one node
phy <- rcoal(10)
node <- 13
tipDes(phy, node)
#> [1] "t3"  "t7"  "t6"  "t1"  "t9"  "t10" "t8" 
## multiple nodes
node <- c(13,14,15)
tipDes(phy, node)
#> $`13`
#> [1] "t3"  "t7"  "t6"  "t1"  "t9"  "t10" "t8" 
#> 
#> $`14`
#> [1] "t3"  "t6"  "t10" "t8" 
#> 
#> $`15`
#> [1] "t7" "t1" "t9"
#>